Improved dimension text processor to be able to handle nested definitions.
Uses a multi group construct with regular expressions that are able to
detect and extract recursive capture groups. Once the larger nested
multi groups are extracted and processed, the remaining single groups
are subsequently incorporated. Certain optional and required
restrictions are imposed on the processing, ie. balanced parenthesis
groups. Overall, the algorithm is a bit hacky and needs to be cleaned up
a bit more.