Major improved of logic in UnitsDataModel related components
Introduced a partially finished implementation in the UnitContentDeterminer component that tries to find the best reduced match for a given unit or component unit dimension using combinaotric partitioning.  Restructured logic in the UnitContentMainRetriever component that tries to find or create a unit instance based on provided text.  Modified misc logic and comments in other UnitsDataModel related components.