MainActivity and AutoComplete Updates
Improved automcomplete to account for the presence of prefixes. Updated the main activity to conditionally format conversion values in cases where the target quantity has multiple units. If the conversion value is a very large or small number, then scientific notation is used. Next, if the conversion value can be represented as a simplified fraction where the denominator is less than 20, then fractional formatting is used. Otherwise, the conversion value is rounded to an adjustable decimal place based on logarithmic scaling.